requirements.txt
streamlit==1.26.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
gspread==5.10.0
pandas==2.0.3
reportlab==4.0.4
pytz==2023.3
streamlit-lottie==0.0.3
Pillow==10.0.0
passlib==1.7.4
bcrypt==4.0.1
